Re-order opening of HTTP Ports
Open the configured HTTP port first (instead of port 80) and indicate the server is active even if opening of port 80 is not successful.  This will allow cetonproxy to be run by a non-root user inside a Wine environment on Linux for use with Emby/Plex.  Port 80 is a privileged port on Linux, so not available to a non-root user without jumping through some hoops.  An alternative would be to include a configuration parameter (or GUI option) to inhibit attempting to opening of port 80 for configurations where it is not needed.